# src/BackEnd/Api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "心之旅",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/appt/list": {
            "post": {
                "tags": [
                    "总台管理系统-预约管理"
                ],
                "summary": "预约列表",
                "operationId": "App\\Http\\Controllers\\Admin\\AppointmentController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "storeId": {
                                        "description": "门店编号",
                                        "type": "string",
                                        "default": "13888888888"
                                    },
                                    "state": {
                                        "description": "状态",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "userId": {
                                        "description": "人员编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "wcId": {
                                        "description": "小程序用户编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "startDate": {
                                        "description": "预约开始时间",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "endDate": {
                                        "description": "预约结束时间",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "pageSize": {
                                        "description": "条数",
                                        "type": "number",
                                        "default": "10"
                                    },
                                    "page": {
                                        "description": "页数",
                                        "type": "number",
                                        "default": "1"
                                    },
                                    "searchKey": {
                                        "description": "关键字",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'page':['\\u8bf7\\u8f93\\u5165\\u5355\\u4f4d\\u540d\\u79f0\\uff01']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/appt/getWorktime": {
            "post": {
                "tags": [
                    "总台管理系统-预约管理"
                ],
                "summary": "人员排班表",
                "operationId": "App\\Http\\Controllers\\Admin\\AppointmentController::GetWorktime",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "storeId"
                                ],
                                "properties": {
                                    "storeId": {
                                        "description": "门店编号",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "workDay": {
                                        "description": "日期",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "workTime": {
                                        "description": "时间",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "userId": {
                                        "description": "用户编号",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取列表失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{'page':['\\u8bf7\\u8f93\\u5165\\u5355\\u4f4d\\u540d\\u79f0\\uff01']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/address/list": {
            "get": {
                "tags": [
                    "小程序-收货地址"
                ],
                "summary": "获取收货地址列表",
                "operationId": "App\\Http\\Controllers\\WechatUser\\AddressController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "获取收货地址列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[{'addrId':2,'name':'xxx','phone':'13828242712','address':'\\u8fbe\\u5a03\\u5927','default':0},{'addrId':3,'name':'xxx1111','phone':'13828242712','address':'\\u8fbe\\u5a03\\u59271111','default':0},{'addrId':4,'name':'xxx1111','phone':'13828242712','address':'\\u8fbe\\u5a03\\u59271111','default':0},{'addrId':6,'name':'xxx1111','phone':'13828242712','address':'\\u8fbe\\u5a03\\u59273333','default':1}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/pet/list": {
            "get": {
                "tags": [
                    "小程序-宠物"
                ],
                "summary": "获取宠物列表",
                "operationId": "App\\Http\\Controllers\\WechatUser\\PetController::GetList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "获取宠物列表成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[{'id':1,'avatar':'11','nickname':'11','breed':'11','gender':'11'}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}